Would this finally be the good version of Resources' Management I dreamed of since I read about AssetReferences and Addressables?
Been dealing for the best way to use AssetReferences as Keys (RuntimeKey didn't work as expected, its API documentation is wrong since it is not a Hash128, but object, so at the end I will just pass AssetReference as Key, it works and allows for single entries). Now I will do full replacement to Moskar Scene using AssetReference and not indices, including Mateo.

It is kinda expected that since AssetReference is a class, it wouldn't be recommended to treat as a Dictionary key. I made tests with string as key (RuntimeKey.ToString from the AssetReference) and it worked, but I felt that was very open to input any key that didn't come from an AssetReference, and my intention was to give type safety and avoid surface conversions, just pass the AssetReference as key and get an Object from the Dictionary.

The solution was to make classes inheriting from AssetReference and override Equals and GetHashCode, 2 essential Object's functions to implement for data to be treated as Dictionary keys.

Implemented:
 - VAssetReference, implementing IEqualityComparer<T> and overriding Equals(object) and GetHashCode()
 - VAssetReferenceT<T>, inheriting from AssetReferenceT<T>, implementing IEqualityComparer<T> and overriding Equals(object) and GetHashCode()

Modified:
 - PoolManager has changed the type of  pools' dictionaries, now the keys are (again) AssetReferences instead of objects. The same treatment has been made for ResourcesManager
 -  VAddressables have methods to release a set of Components and Objects, now passing an IEnumerable<T> instead of a Dictionary, which gives more flexibility.
 - Now GameObjectPool<T> by default creates a first instance, though it is optionally (size parameter of its constructor)
 - All classes and structs made under the namespace Flamingo that have declared an AssetReference is instead changed for VAssetReference

Made a test by instantiating a Moskar from the pool by giving a VAssetReference and it worked.

Now, please let me finish Moskar's Scene...

// Lif